+++ packages/plugin-serverless/src/commands/serverless/list-templates.js
@@ -10,10 +10,12 @@ const {
   normalizeFlags,
 } = require('../../utils');
 
+const { flags, aliasMap } = convertYargsOptionsToOclifFlags(cliInfo.options);
+
 class FunctionsListTemplates extends Command {
   async run() {
     let { flags, args } = this.parse(FunctionsListTemplates);
-    flags = normalizeFlags(flags);
+    flags = normalizeFlags(flags, aliasMap);
 
     const opts = Object.assign({}, flags, args);
     return handler(opts, undefined);
@@ -24,8 +26,6 @@ FunctionsListTemplates.description = describe;
 
 FunctionsListTemplates.args = [];
 
-FunctionsListTemplates.flags = Object.assign(
-  convertYargsOptionsToOclifFlags(cliInfo.options)
-);
+FunctionsListTemplates.flags = Object.assign(flags);
 
 module.exports = FunctionsListTemplates;