+++ packages/runtime-handler/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.0.2](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.0.2-beta.3...@twilio/runtime-handler@1.0.2) (2021-05-19)
+
+**Note:** Version bump only for package @twilio/runtime-handler
+
+
+
+
+
 ## [1.0.2-beta.3](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio/runtime-handler@1.0.2-beta.2...@twilio/runtime-handler@1.0.2-beta.3) (2021-04-21)
 
 **Note:** Version bump only for package @twilio/runtime-handler