+++ packages/serverless-api/CHANGELOG.md
@@ -3,6 +3,57 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [5.0.0](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-api@5.0.0-beta.6...@twilio-labs/serverless-api@5.0.0) (2021-05-19)
+
+**Note:** Version bump only for package @twilio-labs/serverless-api
+
+
+
+
+
+# [5.0.0-beta.6](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-api@5.0.0-beta.5...@twilio-labs/serverless-api@5.0.0-beta.6) (2021-05-17)
+
+
+### Bug Fixes
+
+* **twilio-run:** show service sid if available instead of name ([#265](https://github.com/twilio-labs/serverless-toolkit/issues/265)) ([816a26f](https://github.com/twilio-labs/serverless-toolkit/commit/816a26f7d64ed4193cd02a68abc8667334001771)), closes [#262](https://github.com/twilio-labs/serverless-toolkit/issues/262)
+
+
+### BREAKING CHANGES
+
+* **twilio-run:** The output of the config info in deployment changed
+
+
+
+
+
+# [5.0.0-beta.5](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-api@5.0.0-beta.4...@twilio-labs/serverless-api@5.0.0-beta.5) (2021-05-13)
+
+
+### Bug Fixes
+
+* **serverless-api:** fixes assets paths in deployment ([#258](https://github.com/twilio-labs/serverless-toolkit/issues/258)) ([e37ecb3](https://github.com/twilio-labs/serverless-toolkit/commit/e37ecb336eaec74130d98066a63b858d71f26f62))
+
+
+
+
+
+# [5.0.0-beta.4](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-api@5.0.0-beta.3...@twilio-labs/serverless-api@5.0.0-beta.4) (2021-04-30)
+
+
+### Bug Fixes
+
+* **assets, serverless-api:** uri encode path names for asset versions ([c2b1f3d](https://github.com/twilio-labs/serverless-toolkit/commit/c2b1f3d16ce6af31ef2aaa1d086b3565bbcd1c7d))
+
+
+### Features
+
+* **plugin-assets:** first version of working assets plugin ([4ee08d3](https://github.com/twilio-labs/serverless-toolkit/commit/4ee08d3ebcb82abe1cb11752d45021f2ca3e03c6))
+
+
+
+
+
 # [5.0.0-beta.3](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-api@5.0.0-beta.2...@twilio-labs/serverless-api@5.0.0-beta.3) (2021-04-21)
 
 **Note:** Version bump only for package @twilio-labs/serverless-api