+++ packages/serverless-twilio-runtime/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-twilio-runtime@2.0.0-beta.3...@twilio-labs/serverless-twilio-runtime@2.0.0) (2021-05-19)
+
+**Note:** Version bump only for package @twilio-labs/serverless-twilio-runtime
+
+
+
+
+
 # [2.0.0-beta.3](https://github.com/twilio-labs/serverless-toolkit/compare/@twilio-labs/serverless-twilio-runtime@2.0.0-beta.2...@twilio-labs/serverless-twilio-runtime@2.0.0-beta.3) (2021-04-21)
 
 **Note:** Version bump only for package @twilio-labs/serverless-twilio-runtime